Move the Export test into test_reporter.py
Keeps the reporter tests in the file that already covers ReporterAgent
instead of adding a second module for one test.
